refactor create adventure for reuseability
